cleaned up utils
