+++ client/app/components/WikiEditor/helpers/getLinkEntities.js
@@ -1,0 +1,15 @@
+const getLinkEntities = (contentBlock, callback, contentState) => {
+  contentBlock.findEntityRanges(
+    (character) => {
+      const entityKey = character.getEntity();
+
+      return (
+        entityKey !== null &&
+        contentState.getEntity(entityKey).getType() === 'LINK'
+      );
+    },
+    callback
+  );
+}
+
+export default getLinkEntities;
