+++ client/app/components/WikiEditor/helpers/index.js
@@ -1,0 +1,13 @@
+import getLinkEntities from './getLinkEntities';
+import convertToEditorState from './convertToEditorState';
+import getCurrentEntityKey from './getCurrentEntityKey';
+import shouldHidePlaceholder from './shouldHidePlaceholder';
+import getDisabledLevelButtons from './getDisabledLevelButtons';
+
+export {
+	getLinkEntities,
+	convertToEditorState,
+	getCurrentEntityKey,
+	shouldHidePlaceholder,
+	getDisabledLevelButtons,
+};
