+++ client/app/components/WikiEditor/helpers/shouldHidePlaceholder.js
@@ -1,0 +1,11 @@
+const shouldHidePlaceholder = (contentState) => {
+  if (!contentState.hasText()) {
+    if (contentState.getBlockMap().first().getType() !== 'unstyled') {
+      return true;
+    }
+  }
+
+  return false;
+} 
+
+export default shouldHidePlaceholder;
